# docs/contracts/rustshare-api-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RustShare REST API",
    "description": "Auto-generated OpenAPI specification for the RustShare file-sharing and sync platform.",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/",
      "description": "Current server"
    }
  ],
  "paths": {
...
    "/api/v1/notes": {
      "get": {
        "tags": [
          "Notes"
        ],
        "operationId": "list_notes",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of results",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of notes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/NoteSummary"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Notes"
        ],
        "operationId": "create_note",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateNoteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Note created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateNoteResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/notes/{id}": {
      "get": {
        "tags": [
          "Notes"
        ],
        "operationId": "get_note",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Note ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Note content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetNoteResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Note not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Notes"
        ],
        "operationId": "save_note",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Note ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveNoteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Note saved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SaveNoteResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Note not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Notes"
        ],
        "operationId": "delete_note",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Note ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Note deleted"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Note not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateNoteRequest": {
        "type": "object",
        "properties": {
          "content": {
            "type": [
              "string",
              "null"
            ]
          },
          "parent_folder_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "title": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "CreateNoteResponse": {
        "type": "object",
        "required": [
          "id",
          "name",
          "path",
          "content",
          "metadata",
          "current_version",
          "created_at",
          "modified_at"
        ],
        "properties": {
          "content": {
            "type": "string"
          },
          "created_at": {
            "type": "string"
          },
          "current_version": {
            "type": "integer",
            "format": "int32"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "metadata": {
            "$ref": "#/components/schemas/NoteMetadata"
          },
          "modified_at": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "parent_folder_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "path": {
            "type": "string"
          },
          "public_url": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "description": "Standard error response format.",
        "required": [
          "error"
        ],
        "properties": {
          "details": {
            "type": [
              "string",
              "null"
            ]
          },
          "error": {
            "type": "string"
          }
        }
      },
...
      "GetNoteResponse": {
        "type": "object",
        "required": [
          "id",
          "name",
          "path",
          "content",
          "metadata",
          "current_version",
          "created_at",
          "modified_at"
        ],
        "properties": {
          "content": {
            "type": "string"
          },
          "created_at": {
            "type": "string"
          },
          "current_version": {
            "type": "integer",
            "format": "int32"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "metadata": {
            "$ref": "#/components/schemas/NoteMetadata"
          },
          "modified_at": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "parent_folder_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "path": {
            "type": "string"
          },
          "public_url": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "NoteAttachment": {
        "type": "object",
        "required": [
          "file_id",
          "name",
          "mime_type",
          "size",
          "created_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "file_id": {
            "type": "string",
            "format": "uuid"
          },
          "mime_type": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "size": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "NoteMetadata": {
        "type": "object",
        "description": "Note-specific metadata sidecar schema.",
        "required": [
          "kind",
          "title",
          "visibility",
          "created_at",
          "updated_at",
          "excerpt",
          "mime_type",
          "extension"
        ],
        "properties": {
          "attachments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NoteAttachment"
            }
          },
          "color": {
            "type": [
              "string",
              "null"
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "excerpt": {
            "type": "string"
          },
          "extension": {
            "type": "string"
          },
          "icon": {
            "type": [
              "string",
              "null"
            ]
          },
          "kind": {
            "type": "string"
          },
          "mime_type": {
            "type": "string"
          },
          "pinned": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "public_share_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "title": {
            "type": "string"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "visibility": {
            "$ref": "#/components/schemas/NoteVisibility"
          }
        }
      },
...
      "NoteSummary": {
        "type": "object",
        "description": "Note summary for listings (includes file id but not full content).",
        "required": [
          "id",
          "name",
          "path",
          "metadata",
          "owner_id",
          "current_version",
          "size",
          "created_at",
          "modified_at",
          "attachment_count",
          "drawing_count",
          "export_count"
        ],
        "properties": {
          "attachment_count": {
            "type": "integer",
            "format": "int64"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "current_version": {
            "type": "integer",
            "format": "int32"
          },
          "drawing_count": {
            "type": "integer",
            "format": "int64"
          },
          "export_count": {
            "type": "integer",
            "format": "int64"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "metadata": {
            "$ref": "#/components/schemas/NoteMetadata"
          },
          "modified_at": {
            "type": "string",
            "format": "date-time"
          },
          "name": {
            "type": "string"
          },
          "owner_id": {
            "type": "string",
            "format": "uuid"
          },
          "parent_folder_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "path": {
            "type": "string"
          },
          "size": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "NoteVisibility": {
        "type": "string",
        "enum": [
          "private",
          "public"
        ]
      },
...
      "SaveNoteRequest": {
        "type": "object",
        "required": [
          "content"
        ],
        "properties": {
          "attachments": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "$ref": "#/components/schemas/NoteAttachment"
            }
          },
          "color": {
            "type": [
              "string",
              "null"
            ]
          },
          "content": {
            "type": "string"
          }
        }
      },
      "SaveNoteResponse": {
        "type": "object",
        "required": [
          "id",
          "current_version",
          "modified_at",
          "excerpt"
        ],
        "properties": {
          "current_version": {
            "type": "integer",
            "format": "int32"
          },
          "excerpt": {
            "type": "string"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "modified_at": {
            "type": "string"
          }
        }
      },
...
    }
  },
...
}